Rename tfs utils files
