+++ src/rest_parser.cpp
@@ -19,7 +19,7 @@
 #include <string>
 
 #include "rest_utils.hpp"
-#include "tfs_frontend/tfstensorinfo.hpp"
+#include "tfs_frontend/tfs_utils.hpp"
 
 namespace ovms {
