+++ src/rest_utils.cpp
@@ -24,7 +24,7 @@
 #include "tensorflow_serving/util/json_tensor.h"
 #pragma GCC diagnostic pop
 
-#include "tfs_frontend/tfstensorinfo.hpp"
+#include "tfs_frontend/tfs_utils.hpp"
 #include "timer.hpp"
 
 using tensorflow::DataType;
