+++ src/test/deserialization_tests.cpp
@@ -29,7 +29,7 @@
 #pragma GCC diagnostic pop
 
 #include "../deserialization.hpp"
-#include "../tfs_frontend/tfstensorinfo.hpp"
+#include "../tfs_frontend/tfs_utils.hpp"
 #include "test_utils.hpp"
 
 #include <gmock/gmock-generated-function-mockers.h>
