+++ src/test/dummy/1/mapping.json
@@ -1,0 +1,1 @@
+{"inputs": {"b":"input_tensor"}, "outputs": {"a": "output_tensor"}}
