+++ src/test/serialization_tests.cpp
@@ -29,7 +29,7 @@
 #pragma GCC diagnostic pop
 
 #include "../serialization.hpp"
-#include "../tfs_frontend/tfstensorinfo.hpp"
+#include "../tfs_frontend/tfs_utils.hpp"
 #include "test_utils.hpp"
 
 #include <gmock/gmock-generated-function-mockers.h>
