+++ src/tfs_frontend/tfs_utils.cpp
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //*****************************************************************************
-#include "tfstensorinfo.hpp"
+#include "tfs_utils.hpp"
 
 #include <map>
 #include <memory>
